+++ CHANGELOG.md
@@ -13,6 +13,7 @@ New features
 
 * Add `--delev-random`(short `r`) to delevel to a random level between original and target levels.
 * Make program halt at the end and wait for keypress when run without arguments. That's primarily done to prevent terminal window immediately closing when run in Windows with double-click(thanks to Walkihr for feedback).
+* Skip multipatch.esp(tes3cmd) by default(optional).
 
 Fixes
 
@@ -20,6 +21,7 @@ Fixes
 
 Miscellaneous
 
+* Print suggestion to add `--all-lists` when multipatch.esp is skipped.
 * Add program run configuration output to log file(or to display with very verbose output -vvvv), showing used command-line arguments, non-default options and settings.
 
 ## 0.4.0 (2024-01-19)